Create UMAPTest
This adds a unit test for the UMAP algorithm.

Currently, this test is failing on purpose to show that the current implementation of the UMAP algorithm is not working as expected.

The idea of the test is to create two distinct clusters of points in 3D space, one having 50 points and one having 100 points. Both clusters are concatened into one double array with 150 rows and 3 columns. The test first uses TSNE to reduce the 3 dimensions to 2 dimensions. This works as intendeded. It can be shown that after running TSNE the two distinct clusters in 3d space form two distinct clusters in reduced 2d space.
The same test with UMAP does however fail. First of all, the test, if all 150 rows exist in the UMAP result fails. This according to my limited mathematical understanding of the implemenation seems to be related to the fact that only the largest connected component of the neighbourhood graph is used. The second assumption, i.e. if we see two distinct point clouds in reduced space cannot be tested.